all tests pass except test_server search
* build/rules.make (all): add smm-server
  (test_all_harness.out): needs smm-server, not smm

* source/smm-database.adb (Write_Play_Before_After): new

* source/smm-id3.adb (Read Stream): new, factored out
  (Read file): use it

* source/smm-id3.ads (Read Stream): make visible for unit test

* source/smm-server.adb (Server): make source_root absolute, as required by smm.relative

* source/smm-song_lists.adb (Least_Recent_Songs): fix comment

* test/smm-id3-aunit.adb: New file.
* test/smm-id3-aunit.ads: New file.

* test/test_all_harness.good_out: match test changes, add test_id3, test_server.test_search_initial


* test/test_first_pass_with_last.adb:
* test/test_first_pass_no_last.adb (Name):
* test/test_copy.adb (Name): fix

* test/test_download.adb (Nominal): use only integer download times

* test/test_least_recent.adb (Create_Test_DB): use test_utils.cleanup
  (Check): pass DB
  (Nominal): declare DB locally, so it is finalized when test completes

* test/test_play_before.adb (): change to smm.database

* test/test_server.adb (Test_Playlist): match gnat gpl 2017 random
  (Test_Get_File, Set_Up_Case): use smm.id3, smm.database

* test/test_utils.adb (Cleanup): simplify
  (Insert): handle arbitrary 1 digit times

+++ test/test_utils.adb
@@ -26,11 +26,9 @@ package body Test_Utils is
 
    procedure Cleanup
    is begin
-      Ada.Directories.Delete_Tree ("tmp");
-   exception
-   when Ada.Text_IO.Name_Error =>
-      --  already deleted
-      null;
+      if Ada.Directories.Exists ("tmp") then
+         Ada.Directories.Delete_Tree ("tmp");
+      end if;
    end Cleanup;
 
    procedure Create_Empty_DB (DB_File_Name : in String)
@@ -72,7 +70,7 @@ package body Test_Utils is
       Last_Downloaded : in Duration;
       Category        : in String := "vocal")
    is
-      One_Second_Time_String : constant SMM.Database.Time_String := "1958-01-01 00:00:01";
+      Prefix : constant String := "1958-01-01 00:00:0";
    begin
       DB.Insert
         (ID              => ID,
@@ -81,10 +79,7 @@ package body Test_Utils is
          Artist          => "none",
          Album           => "none",
          Title           => "none",
-         Last_Downloaded =>
-           (if Last_Downloaded = 0.0
-            then SMM.Database.Default_Time_String
-            else One_Second_Time_String));
+         Last_Downloaded => Prefix & Duration'Image (Last_Downloaded) (2 .. 2));
    end Insert;
 
    function "&" (Left, Right : in SMM.ID3.Tag) return SMM.ID3.Tag_Lists.List
